+++ mxops/data/__init__.py
@@ -3,3 +3,7 @@ author: Etienne Wallet
 
 This subpackage is used to host the logic for analysing on-chain contracts
 """
+
+from mxops.data.path_versions import v1_0_0 as data_path
+
+__all__ = ["data_path"]